I gave support for Windows directory separators and changed the directory filtering method
Added the '_convertPathSep' method that will check the path.sep looking for a Windows-like directory separator,
and if positive, it will replace unix-like separators for "\\".

Changed the directory filtering logic: because the previous logic was made with a  string match, it not always worked as expected. I have made a string comparision that will check the directory hierarchy.

Walker can be called using paths (also in filters) with unix-like directory separators in Windows and it will work